add "Adafruit ST7735 and ST7789" lib to the project
PlatformIO has issues building the project on Windows when Adafruit ST7735 lib
is set as lib deps from registry. It tries to install some 3rd party SD lib an
eventually failes to build.
This does NOT reproduces on linux somehow :)
Include the lib in project for now unless some other solution is found